+++ src/main/java/models/igp/ospf/OSPFPath.java
@@ -1,3 +1,4 @@
+package models.igp.ospf;
 import java.util.ArrayList;
 
 public class OSPFPath {
