+++ src/main/java/models/igp/ospf/OSPFShortestPathTree.java
@@ -1,3 +1,4 @@
+package models.igp.ospf;
 import java.util.*;
