+++ src/test/java/models/igp/ospf/OSPFAreaTest.java
@@ -1,4 +1,10 @@
+package models.igp.ospf;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
+
+import models.igp.ospf.OSPFPath;
+import models.igp.ospf.OSPFShortestPathTree;
+
 import com.fasterxml.jackson.core.type.TypeReference;
 import java.io.File;
 import java.io.IOException;
